Update patch set 6
Patch Set 6:

BTW, once this is landed, people may think ninja's progress report is broken. The number of total build targets will decrease during the build when the restat feature finds unnecessary rebuilds. For the case I wrote in the commit message, the ninja shows XX/2563 at first and then the denominator will drop to 90. I'm not sure if this change is big enough and we need an announcement.

Patch-set: 6