Explicitly use messaging app in SDK phone builds
This change explicitly includes only the messaging app instead of all of
aosp_base_telephony. This fixes the issue where on initial boot the user
was prompted to select a Home app. This happened because the
aosp_base_telephony makefile includes other makefiles that in turn add
the package Provision which has an activity with an intent filter for
handling HOME. SdkSetup already had such an activity so with two choices
it was left to the user to decide which one to use.

BUG=27273457

Change-Id: I42e77b892ce235a5fa858561325ae37973cfb0cd
(cherry picked from commit 12de252f451b34cfa67f7256027dad84dcfd984d)