Update patch set 2
Patch Set 2:

There is plenty of code directly using arm_neon.h in the tree, and even compiler-inserted headers are supposed to be in the header dependency dump.

The intel version of arm_neon.h in the gcc host toolchain appears to be custom to support x86, the clang host toolchain only has the arm-compatible version generated with tblgen.

Patch-set: 2
Label: Verified=0