Update patch set 3
Patch Set 3:

> > This new utility shares quite some code with dexdump indeed. To
 > > avoid the maintenance nightmare of having more utilities than we
 > > need, how would you feel about making this a switch in dexdump?
 > >
 > > This way, minor bug fixes in the driver and related code don't
 > need
 > > to be repeated in several similar files. Also, it will make the
 > > libart port to windows easier.
 > 
 > In that respect I was wondering before: can a simple dexdump output
 > actually be parsed to do the same job?

Thanks for comments! I'll move this to art/ and reply to inlined comments.

Re dexdump: I introduced a new tool based on bdc's comment: https://b.corp.google.com/u/0/issues/24597504#comment7

It seems there're pros-and-cons. As Aart mentioned having one more tool would be a maintenance burden. I think the benefit of having a separate tool is we'll unlikely regress the output due to future refactoring for dexdump. We'd able to mitigate this by checking in prebuilt dexdump if we want. I was also feeling dexdump's code has already had quite a few branches especially for the output format (plain vs XML), but this seems not to be a big deal. I'm fine with either way.

Patch-set: 3