Default LOCAL_MULTILIB:=both for HOST_NATIVE_TEST
The target BUILD_NATIVE_TEST defaults to building both 32 and 64-bit
versions of the tests. Many of our host tests have opted into this
manually, but some have not. Switch the default on the host tests to
stay consistent.

To preserve script compatibility, a symlink remains pointing to the
binary that would have been built before this change.

Change-Id: I9585984e933c2ac36134fc1968736c0bb5fea7be