Update patch set 1
Patch Set 1:

libselinux and liblog are the only worrisome set that doing a reverse name to tag number lookup at runtime could turn into an issue where hard-coding into a header file would be a clear advantage. The executables would not be under the same time-crunch to check the values in the runtime list.

Patch-set: 1