Allow smooth upgrade of AVD's from older API's.
If we are using an old API for the emulator
(API 19 for example) and then upgrade the AVD
to API 23 through the AVD settings, we get a bunch
of crashes in the TelephonyProvider,
as it is looking for a file that does not exist
that concerns upgrading from old "apn"'s.

This CL adds this file to the files to be copied
to /etc in the system image.

BUG: b.android.com/200390
BUG=27273457

RELEASE NOTE: Fixed telephony service crashes on startup when upgrading from an old API.

Change-Id: I227713880bf20541c759aab588c00197ac7aebfd
(cherry picked from commit e5083159401edf1df5109b56d1aedbbb739584e6)