Update patch set 2
Patch Set 2:

> > Sorry, it's sorted list. But subtraction, union, interset etc is
 > > using a priority queue and expects sorted input. Right now, I
 > > simply makes two copies of rangeset. The file_map stores the
 > > unsorted one, and all operations uses the sorted one. It make
 > sense
 > > to have the unsorted copy as internal storage if rangeset.
 > 
 > i don't get it... are you gonna upload a new patch set to address
 > the issues in subtract/union?

a clarification to my concern: a value in file_map might be used in subtraction, so we need to make sure it computes the correct result. think about the case that file_map['abc.apk'] = "15-20 30 1-5", and some blocks (say "30") are stashed in earlier transfers. so the source blocks for this transfer needs to be "15-20 1-5". in the current PS, i don't think this case is well handled.

and my proposed solution is to keep the original input string for this RangeSet if order matters. internally it still uses the sorted tuples (1, 6, 30, 31, 15, 21) which support subtraction, and we pass along the original input string into the subtraction result. when touching order-sensitive parts like ReadRangeSet() or RangeSet.to_string(), we output according to the original input string order.

Patch-set: 2