Update patch set 1
Patch Set 1: Code-Review+1

lgtm

Thoughts:

In order to keep group and passwd sanity, we will likely need both files. group will have the first three fields identical content as passwd.

The defaults are perfectly fine. We will restrict the valid UID/GIDs to the oem range when we read from these files. I am thinking that the gid field in the passwd file will be generated the same as the uid field; but we may consider reading it as we will for the home directory and shell when reporting to the caller _even_ if it is different. If it is blank, we will set the gid to the same as the uid on report back, so you could leave it empty?

Given the default behavior in our planned reader, it is possible that if only three fields are recorded in passwd, that the group and passwd can be made identical (one file, and a hard-link). I am unsure if breaking that far away from POSIX format is worth the file-space and mmap'ing/buffer cache savings when reading?

Patch-set: 1
Label: Code-Review=+1