Indicate that all emulator images support USB
The Android framework assumes that if an app uses UsbManager then there
must be USB support on the device. In order to deal with this the
emulator has to indicate USB support. Previously this was true for
Google SDK phone images because USB support is part of the common API
packages that gets installed. For AOSP and some other build targets USB
support was not included. This change indicates support for USB for all
emulator system images.

BUG: 28088069

Change-Id: I880210e89e6ad3412bdc069d6c20ed54a52cff83