Remove Romansh from PRODUCT_LOCALES
The inclusion of Romansh (rm_CH), while it did not have localized
strings, was causing the incorrect assumption that the system
supports Romansh, causing the language fallback to work incorrectly
when Romansh was in the list of user's locales.

Hindi (hi_IN) was also included twice, and this removes the duplicate
too.

Bug: 27325465
Change-Id: I87ec06dc459137f71bb8e5e2f431f629bd51928e