Update patch set 1
Patch Set 1:

(1 comment)

> I am ok with this argument approach.
 > I was thinking about exporting an enviromental variable from
 > vendor/google/vendorsetup.sh, which is included by
 > build/envsetup.sh.

One thing I think this approach isn't good for is when the URL is a collection of git repositories rather than the regular source tree. It would be fairly easy to add mappings, etc. but it'd be nice to reuse readily available information, such as does the directory contain a .git directory. Looking in Google3, it wasn't clear to me this script was always run after envsetup.sh.

Patch-set: 1