Remove Dev Settings app from emulator images
The Dev Settings app currently just launches an intent that is received
by the regular Settings app. However, since N the Settings app only
receives the intent when developer mode is enabled. If it is not enabled
the Dev Settings app will crash. Since the Dev Settings app doesn't
really add much benefit this change removes it from the emulator system
images to prevent this issue. The user will have to access the developer
settings from the regular Settings app just like on a physical device.

BUG: 27549126

Change-Id: I8191f9003bb452ef76c4b236b965368b48be3d25