Update patch set 7
Patch Set 7:

> This still requires tools that are not available to use by the build system, doesn't it?

It requires the make_efi_image program which was added in https://android-review.googlesource.com/#/c/243590/

Right now make_efi_image is indeed using tools from the host (mkfs.vat, mtools) but this might change in the future. For now this tool is only used by the brillo_uefi_x86_64 board and for this board it's noted in the README that these tools need to be installed.

Patch-set: 7