Update patch set 3
Patch Set 3:

Yes, we currently require a rebuild of all GOMA-enabled modules when switching GOMA off (or on, but that's less important).

We don't want to rebuild everything if we can avoid it -- goma only applies to ~1/3 of the commands we run. And we don't have any way to do a global rebuild on environment variable change, since the environment is not tracked by ninja.

The right way to do this would be to set GOMACC_PATH individually for each command that uses it, then the existing ninja functionality will re-run those commands.

Patch-set: 3