Update patch set 1
Patch Set 1:

It's just because I'd like to do a simple strings.Join(<slice>, "|") similar to the comma-separated DEFAULT_GLOBAL_TIDY_CHECKS. So it generates the same list as make, just without the spaces that the make version added. For DEFAULT_LOCAL_TIDY_CHECKS, it's a custom structure, so I'm using a dedicated function to convert it over to the make syntax.

(I'll upload the go side of this with the rest of the clang-tidy functionality, which I haven't finished yet)

Patch-set: 1