Update patch set 4
Patch Set 4:

> As per the TreeHugger output, before submitting, do I need to fixup
 > these branches:
 > 
 > git_master/bullhead-userdebug_fastbuild_linux
 > git_master/angler-userdebug_fastbuild_linux
 > git_master/gce_x86_phone-userdebug_fastbuild_linux
 > 
 > Or do I submit now and fixup the branches later?

I'm not sure about the best practice. I usually submit a cherrypick into the broken branch first (in this case, goog/master). After that one lands, let the TreeHugger retry. In the worst case if it's still broken, I just submit anyway and choose to skip this one in the auto merger (since it's been merged).

Patch-set: 4