Update patch set 1
Patch Set 1:

What configurations are warning here? I don't see any warnings on my local builds.

Zero isn't a valid value that we should be putting into the entry, so I'm worried that we're not really initializing these elsewhere. (Or this is a false-positive, and this just works around a problem with the warning)

Patch-set: 1
Label: Verified=0