Update patch set 2
Patch Set 2:

> Losing sync with the issues, is there a bug with a story describing
 > the regressions and the causes?

my assumption is that http://b/33535908 is the real issue. there was a bionic test failures bug (caused by the bogus AID_APP and AID_USER entries), and an app install time regression duped into the bug above that i'm assuming was just a side-effect of the logspam using up a ton of cpu.

my problem is that even today, on internal master, with marlin, i can't reproduce the problem. i've reapplied the system/core patch, the bionic patch, and this, but i don't see the logspam. were you able to reproduce it?

Patch-set: 2
Label: Verified=0